Move contract testing to lib/
