+++ lib/src/testing/__examples__/default-response.ts
