+++ lib/src/testing/__examples__/single-provider-state.ts
