+++ lib/src/testing/test-timer.ts
